requirements.txt
streamlit
pandas
openai
python-dotenv
pytz
python-dateutil
azure-core
azure-identity
azure-storage-blob
requests
pandas
matplotlib # Add this if you are doing any plotting
 